Move CompatibilityIssuesWarning to dumbs
