+++ src/custom/pages/Swap/SwapMod.tsx
@@ -43,7 +43,7 @@ import CowSubsidyModal from 'components/CowSubsidyModal'
 import { AlertWrapper } from './styleds'
 import { setMaxSellTokensAnalytics } from 'utils/analytics'
 import { ImportTokenModal } from 'cow-react/swap/smart/ImportTokenModal'
-import { CompatibilityIssuesWarning } from 'cow-react/swap/smart/CompatibilityIssuesWarning'
+import { CompatibilityIssuesWarning } from 'cow-react/swap/dumb/CompatibilityIssuesWarning'
 import { ConfirmSwapModalSetup, ConfirmSwapModalSetupProps } from 'cow-react/swap/smart/ConfirmSwapModalSetup'
 import { useAtomValue } from 'jotai/utils'
 import { swapConfirmAtom } from 'cow-react/swap/state/swapConfirmAtom'
